implementation of basic wasm pusher
hardcoded sdk config
hardcoded peer communication
hardcoded file with wasm

takes a wasm file, hex encodes it, and submits it as a transaction

ledger client was old code from a different program and was useful for debugging.

ought to read most of these args from cli or environment

prints the raw response to the console, which contains raw protobuf output.

resolves #3



Signed-off-by: shubham aggarwal <shubham6300@gmail.com>